requirements.txt
opencv_python
Flask==1.0.2
Pillow==6.0.0
torchvision==0.2.1
numpy==1.15.4
gunicorn==19.9.0
https://download.pytorch.org/whl/cpu/torch-0.4.1-cp36-cp36m-linux_x86_64.whl
